refactor : HealthMetricChartController 제거 및 엔드포인트 변경
